Fixed arrays. Added primitive wrappers. Added putfield and getfield. PrintStream now accepts an Object. Fixed array depth bug in parameter parsing. Probably a bunch of other stuff.

+++ CCLib/src/java/lang/Float.java
@@ -1,9 +1,9 @@
 package java.lang;
 
-public class Long extends Number {
-    private final long value;
+public class Float extends Number {
+    private final float value;
 
-    public Long(long value) {
+    public Float(float value) {
         this.value = value;
     }
